requirements-ocr.txt
fastapi
uvicorn[standard]
python-multipart
requests
google-generativeai
